+++ src/JokeHeader.js
@@ -4,7 +4,7 @@ class JokeHeader extends Component {
   render() {
     return (
     <div className="appHead">
-      <h2 className="dadJokesHead">Dad Jokes!</h2>
+      <h2 className="dadJokesHead">The Joke Factory</h2>
     </div>
     )
   }
